Deprecate passing extras with graphs
