+++ pytorch_lightning/trainer/connectors/logger_connector/result.py
@@ -26,10 +26,13 @@ from pytorch_lightning.utilities.device_dtype_mixin import DeviceDtypeModuleMixi
 from pytorch_lightning.utilities.enums import LightningEnum
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.metrics import metrics_to_scalars
+from pytorch_lightning.utilities.warnings import WarningCache
 
 # re-define the ones from pytorch_lightning.utilities.types without the `Number` type
 _METRIC = Union[Metric, torch.Tensor]
 _METRIC_COLLECTION = Union[_METRIC, Mapping[str, _METRIC]]
+
+warning_cache = WarningCache()
 
 
 class MetricSource(LightningEnum):
@@ -279,9 +282,16 @@ class ResultCollection(dict):
 
         def check_fn(v):
             if v.grad_fn is not None:
-                raise MisconfigurationException(f'You returned a tensor with `grad_fn`. The extra values are {extra}')
+                # raise MisconfigurationException(f'You returned a tensor with `grad_fn`. The extra values are {extra}')
+                warning_cache.warn(
+                    f"One of the returned values {set(extra.keys())} has a `grad_fn`. We will detach it automatically"
+                    " but this behaviour will change in v1.6. Please detach it manually:"
+                    " `return {'loss': ..., 'something': something.detach()}`", DeprecationWarning
+                )
+                return v.detach()
+            return v
 
-        apply_to_collection(extra, torch.Tensor, check_fn)
+        extra = apply_to_collection(extra, torch.Tensor, check_fn)
         self['_extra'] = extra
 
     def log(
