+++ tests/trainer/logging_/test_train_loop_logging.py
@@ -690,16 +690,16 @@ def test_logging_raises(tmpdir):
     with pytest.raises(MisconfigurationException, match='`self.log` with the key `foo/dataloader_idx_0`'):
         trainer.fit(model)
 
-    class TestModel(BoringModel):
+    # class TestModel(BoringModel):
 
-        def training_step(self, *args):
-            loss = super().training_step(*args)['loss']
-            return {"loss": loss, 'foo': loss}
+    #     def training_step(self, *args):
+    #         loss = super().training_step(*args)['loss']
+    #         return {"loss": loss, 'foo': loss}
 
-    trainer = Trainer(default_root_dir=tmpdir)
-    model = TestModel()
-    with pytest.raises(MisconfigurationException, match='You returned a tensor with `grad_fn`'):
-        trainer.fit(model)
+    # trainer = Trainer(default_root_dir=tmpdir)
+    # model = TestModel()
+    # with pytest.raises(MisconfigurationException, match='You returned a tensor with `grad_fn`'):
+    #     trainer.fit(model)
 
     class TestModel(BoringModel):
